tests(helpers): separate db functions
